Fix error on inheritance between base model and drivers
