+++ src/source_drivers/modbus/models/device.py
@@ -1,12 +1,11 @@
 from sqlalchemy.orm import validates
 
 from src import db
+from src.models.device.model_device_mixin import DeviceMixinModel
 from src.source_drivers.modbus.models.network import ModbusType
-from src.models.device.model_driver_device import DriverDeviceModel
 
 
-class ModbusDeviceModel(DriverDeviceModel):
-    DRIVER_NAME = 'Modbus'
+class ModbusDeviceModel(DeviceMixinModel):
     __tablename__ = 'modbus_devices'
 
     type = db.Column(db.Enum(ModbusType), nullable=False)
@@ -20,6 +19,10 @@ class ModbusDeviceModel(DriverDeviceModel):
     timeout_global = db.Column(db.Boolean(), nullable=False)
     last_poll_timestamp = db.Column(db.DateTime)
     fault_timestamp = db.Column(db.DateTime)
+
+    @classmethod
+    def get_polymorphic_identity(cls):
+        return 'Modbus'
 
     @validates('tcp_ip')
     def validate_tcp_ip(self, _, value):
