+++ src/source_drivers/modbus/models/point.py
@@ -1,10 +1,10 @@
 from src import db
-from src.models.point.model_driver_point import DriverPointModel
+
+from src.models.point.model_point_mixin import PointMixinModel
 from src.source_drivers.modbus.interfaces.point.points import ModbusPointType, ModbusDataType, ModbusDataEndian
 
 
-class ModbusPointModel(DriverPointModel):
-    DRIVER_NAME = 'Modbus'
+class ModbusPointModel(PointMixinModel):
     __tablename__ = 'modbus_points'
 
     reg = db.Column(db.Integer(), nullable=False)
@@ -17,3 +17,7 @@ class ModbusPointModel(DriverPointModel):
     data_offset = db.Column(db.String(80), nullable=False)
     timeout = db.Column(db.Float(), nullable=False)
     timeout_global = db.Column(db.Boolean(), nullable=False)
+
+    @classmethod
+    def get_polymorphic_identity(cls):
+        return "Modbus"
